% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSumClustsDPM.R
\name{checkConvergence}
\alias{checkConvergence}
\title{checkConvergence}
\usage{
checkConvergence(PSMs)
}
\arguments{
\item{PSMs}{3-dimensional array of PSMs, for each j PSMs[,,j] is the PSM of subsampled chain j, j = 1, ..., m}
}
\value{
results: list of length m-1 of results of processPSM function applied to the following subsets of PSMs: 1,2; 1,2,3; 1,2,3,4; ... 1,2,...,m

coph: vector of length m-1 of cophenetic correlation coefficients measuring how well the summary PSM obtained using the PY+PEAR method on the first
k submatrices (k=1, ..., l) is described by a hierarchcial clustering tree for a summary PSM based on the first l subsampled chains

coph_DP: as coph, but for summary PSMs obtained by the DPM+PEAR method

distPY: vector of length m-2 of Frobenius (Euclidean) norm of distances between consecutive summary PSMs, that is summary PSMs obtained from chains 1,2,..,k,k+1 and 1,2,...k (PY+PEAR)

distDP: as distPY, but for DPM+PEAR method
}
\description{
checkConvergence
}
\author{
Magdalena Strauss
}
