% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSumClustsDPM.R
\name{computeSumClustPEAR}
\alias{computeSumClustPEAR}
\title{computeSumClustPear}
\usage{
computeSumClustPEAR(PSM, maxCl = 10)
}
\arguments{
\item{PSM}{posterior similarity matrix}
}
\value{
Summary clustering computed using the PEAR criterion (Fritsch and Ickstadt, 2009, using the mcclust package
 (Fritsch, 2012))
}
\description{
Internal function
}
\author{
Magdalena Strauss
}
