% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSumClustsDPM.R
\name{processPSMs}
\alias{processPSMs}
\title{processPSMs}
\usage{
processPSMs(PSMs)
}
\arguments{
\item{PSMs}{3-dimensional array of PSMs, for each j PSMs[,,j] is the PSM of subsampled chain j}
}
\value{
weightedPSM: weighted summary PSM obtained using a Pitman-Yor process mixture model
with variable selection

sumClustPEAR: final summary clustering obtained from weightedPSM using the PEAR criterion

weightedPSM_DP weighted summary PSM obtained using a Dirichlet process mixture model
with variable selection

sumClustPEAR_DP: final summary clustering obtained from weightedPSM_DP using the PEAR criterion

weights: weights which were used for the computation of the summary PSM (Pitman-Yor based model)

weights_DP: weights which were used for the computation of the summary PSM (Dirichlet based model)
}
\description{
processPSMs
}
\author{
Magdalena Strauss
}
