% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineSumClustsDPM.R
\name{computeSumClustPEAR}
\alias{computeSumClustPEAR}
\title{computeSumClustPear}
\usage{
computeSumClustPEAR(PSM, maxCl = 10)
}
\arguments{
\item{PSM}{posterior similarity matrix}
}
\value{
Summary clustering computed using the PEAR criterion (Fritsch and Ickstadt, 2009, using the mcclust package
 (Fritsch, 2012))
}
\description{
Compute summary PSMs (posterior similarity matrices) from a set of multiple PSMs obtained for instance
by means of subsampling. This implements the Dirichlet process and Pitman-Yor process based methods
for combining PSMs proposed in Strauss et al. Unravelling shared pseudo-trajectories at single-cell resolution.
Internal function
}
\author{
Magdalena Strauss
}
